Freshen gene annotations using Bioconductor annotation data

Freshen gene annotations using Bioconductor annotation data

freshenGenes(
  x,
  annLib = c("org.Hs.eg.db"),
  tryList = c("SYMBOL2EG", "ACCNUM2EG", "ALIAS2EG"),
  finalList = c("SYMBOL"),
  split = "[, /]+",
  sep = ",",
  handle_multiple = c("first_try", "first_hit", "all"),
  empty_rule = c("original", "empty"),
  include_source = FALSE,
  verbose = FALSE,
  ...
)

Arguments

x
character vector or data.frame with one or most columns containing gene symbols.

annLib
character value indicating the name of the Bioconductor annotation library to use when looking up gene nomenclature.

tryList
character vector indicating one or more names of annotations to use for the input gene symbols in x. The annotation should typically return the Entrez gene ID, usually given by '2EG' at the end of the name. For example SYMBOL2EG will be used with annLib "org.Hs.eg.db" to produce annotation name "org.Hs.egSYMBOL2EG". Note that when the '2EG' form of annotation does not exist, it will be derived using AnnotationDbi::revmap(). For example if "org.Hs.egALIAS" exists, but not "org.Hs.egALIAS2EG", then this function will create a reverse-mapped "org.Hs.egALIAS2EG" derived from "org.Hs.egALIAS".

finalList
character vector to use for the final conversion step. When finalList is NULL no conversion is performed. When finalList contains multiple values, each value is returned in the output. For example, finalList=c("SYMBOL","GENENAME") will return a column "SYMBOL" and a column "GENENAME".

sep
character value used to separate delimited values in x. For example when sep="," then comma-delimited values will be split into separate columns, and each column can be used in the gene annotation update. See handle_multiple.

handle_multiple
character value indicating how to handle multiple values: "first_hit" will query each column of x until it finds the first possible returning match, and will ignore all subsequent possible matches for that row in x. For example, if one row in x contains multiple values, only the first match will be used. "first_try" will return the first match from tryList for all columns in x that contain a match. For example, if one row in x contains two values, the first match from tryList using one or both columns in x will be maintained. Subsequent entries in tryList will not be attempted for rows that already have a match. "all" will return all possible matches for all entries in x using all items in tryList.

empty_rule
character value indicating how to handle entries which did not have a match, and are therefore empty: "original" will use the original entry as the output field; "empty" will leave the entry blank.

include_source
logical indicating whether to include a column that shows the colname and source matched. For example, if column "original_gene" matched "SYMBOL2EG" in "org.Hs.eg.db" there will be a column "found_source" with value "original_gene.org.Hs.egSYMBOL2EG".

verbose
logical indicating whether to print verbose output.

Value

data.frame with one or more columns indicating the input data, then a column "intermediate" containing the Entrez gene ID that was matched, then one column for each item in finalList, by default "SYMBOL".

Details

This function takes a vector or data.frame of gene symbols, and uses Bioconductor annotation methods to find the most current official gene symbol.

The function can also be customized to return additional annotation, or custom annotation.

The annotation process is intended to run in two steps:

Convert the input gene symbol to the recognized Entrez gene ID, usually using something like "org.Hs.egSYMBOL2EG" which returns a character string with Entrez gene identifier. When the gene symbol is not recognized, the backup plan is to use gene aliases, accession numbers, or other relevant annotations to find the current official gene. However, these steps are not as trusted as using the "SYMBOL2EG" approach, so that step is attempted first.

Convert the Entrez gene identifier to the official Entrez gene symbol, usually using something like "org.Hs.egSYMBOL". This step is also a convenient time to include things like the descriptive gene name, which is provided by "org.Hs.egGENENAME".

See also

Examples

if (suppressPackageStartupMessages(require(org.Hs.eg.db))) {
   freshenGenes(c("APOE", "CCN2", "CTGF"));

   ## Optionally show the annotation source matched
   freshenGenes(c("APOE", "CCN2", "CTGF"), include_source=TRUE)

   ## Show comma-delimited genes
   freshenGenes(c("APOE", "CCN2", "CTGF", "CCN2,CTGF"));

   ## Optionally include more than SYMBOL in the output
   freshenGenes(c("APOE", "CCN2", "CTGF"),
      finalList=c("SYMBOL", "ALIAS", "GENENAME"))
}

#>   input intermediate SYMBOL                           ALIAS
#> 1  APOE          348   APOE AD2,APO-E,APOE,ApoE4,LDLCQ5,LPG
#> 2  CCN2         1490   CCN2     CCN2,CTGF,HCS24,IGFBP8,NOV2
#> 3  CTGF         1490   CCN2     CCN2,CTGF,HCS24,IGFBP8,NOV2
#>                                  GENENAME
#> 1                        apolipoprotein E
#> 2 cellular communication network factor 2
#> 3 cellular communication network factor 2