Get annotation database or environment

Get annotation database or environment

get_anno_db(
  x,
  revmap_suffix = c("2EG", "2ENTREZID", "2NAME"),
  ignore.case = FALSE,
  verbose = FALSE,
  ...
)

Arguments

x
character name of an annotation object, or an annotation object itself.

revmap_suffix
character string indicting the expected suffix that can be used to create reverse-mapped annotation data, for example the suffix "2EG" is used to indicate that annotation returns Entrez gene. When annotation does not contain this suffix, the annotation is reverse-mapped using AnnotationDbi::revmap().

ignore.case
logical indicating whether to return an environment after converting all keys to lowercase, which is one implementation choice to provide case-insensitive output from mget(). In order to fulfill the potential, the subsequent mget() must also use tolower(x) on the input character vector. Note that this option is currently fairly slow, and uses more memory while the environment is loaded.

verbose
logical indicating whetheer to print verbose output.

...
additional arguments are ignored.

Details

This function is a simple wrapper function that takes either an annotation data name, for example from the AnnotationDbi package, or an annotation object, and returns the annotation object.

In the event the annotation object must be derived using AnnotationDbi::revmap(), then that process is performed, and the reverse mapped annotation object is returned.

See also